+++ PedalProFrontEnd/PedalProFrontEnd/src/app/BikePart/bicycle-part/bicycle-part.component.ts
@@ -4,6 +4,9 @@ import { Router } from '@angular/router';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { map, Observable, Subject } from 'rxjs';
 import { BicyclePart } from '../../Models/bicycle-part';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent,MyDialogData } from 'src/app/Dialogs/delete-dialog/delete-dialog.component';
+import { ErrorDialogComponent } from 'src/app/Dialogs/error-dialog/error-dialog.component';
 
 @Component({
   selector: 'app-bicycle-part',
@@ -11,12 +14,18 @@ import { BicyclePart } from '../../Models/bicycle-part';
   styleUrls: ['./bicycle-part.component.css']
 })
 export class BicyclePartComponent implements OnInit{
-  constructor(private service:PedalProServiceService,private router:Router, private http:HttpClient){}
+  constructor(private dialog:MatDialog,private service:PedalProServiceService,private router:Router, private http:HttpClient){}
 
   bicycleParts:BicyclePart[]=[];
 
   ngOnInit(): void {
     this.GetBikeParts();
+  }
+
+  openErrorDialog(errorMessage: string): void {
+    this.dialog.open(ErrorDialogComponent, {
+      data: { message: errorMessage }
+    });
   }
 
   //Get bicycle parts
@@ -34,17 +43,34 @@ export class BicyclePartComponent implements OnInit{
   //Delete Bicycle
   DeleteBikePart(id:any)
   {
-    this.service.DeleteBicyclePart(id).subscribe({
-      next:(response)=>{
-        
-        const index=this.bicycleParts.findIndex((bicyclePart)=>bicyclePart.bicyclePartId===id);
-        if(index!=-1){
-          this.bicycleParts.slice(index,1);
-        }
-        this.openModal();
-        
+    const dialogData: MyDialogData = {
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this part?'
+    };
+
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: dialogData
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.service.DeleteBicyclePart(id).subscribe({
+          next:(response)=>{
+            
+            const index=this.bicycleParts.findIndex((bicyclePart)=>bicyclePart.bicyclePartId===id);
+            if(index!=-1){
+              this.bicycleParts.slice(index,1);
+            }
+            this.openModal();
+          },
+          error:(err)=> {
+            const errorMessage = err.error || 'An error occurred';
+            this.openErrorDialog(errorMessage);
+          }
+        })
       }
-    })
+    });
+    
+    
   }
 
   //Refresh the page
