+++ PedalProFrontEnd/PedalProFrontEnd/src/app/ClientType/client-type/client-type.component.ts
@@ -4,6 +4,9 @@ import { Router } from '@angular/router';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { map, Observable, Subject } from 'rxjs';
 import { ClientType } from '../../Models/client-type';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent,MyDialogData } from 'src/app/Dialogs/delete-dialog/delete-dialog.component';
+import { ErrorDialogComponent } from 'src/app/Dialogs/error-dialog/error-dialog.component';
 
 @Component({
   selector: 'app-client-type',
@@ -11,7 +14,7 @@ import { ClientType } from '../../Models/client-type';
   styleUrls: ['./client-type.component.css']
 })
 export class ClientTypeComponent implements OnInit{
-  constructor(private service:PedalProServiceService,private router:Router, private http:HttpClient){}
+  constructor(private service:PedalProServiceService,private dialog:MatDialog,private router:Router, private http:HttpClient){}
 
   clientTypes:ClientType[]=[];
 
@@ -32,22 +35,47 @@ export class ClientTypeComponent implements OnInit{
 // delete method
   DeleteEmpType(id:any)
   {
-    this.service.DeleteClientType(id).subscribe({
-      next:(response)=>{
-        
-        const index=this.clientTypes.findIndex((clientType)=>clientType.clientTypeId===id);
-        if(index!=-1){
-          this.clientTypes.slice(index,1);
-        }
-        this.openModal();
-        
+
+    const dialogData: MyDialogData = {
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this client type?'
+    };
+
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.service.DeleteClientType(id).subscribe({
+          next:(response)=>{
+            
+            const index=this.clientTypes.findIndex((clientType)=>clientType.clientTypeId===id);
+            if(index!=-1){
+              this.clientTypes.slice(index,1);
+            }
+            this.openModal();
+            
+          },
+          error:(err)=>{
+            const errorMessage = err.error || 'An error occurred';
+            this.openErrorDialog(errorMessage);
+          }
+        })
       }
-    })
+    });
+    
   }
 
   ReloadPage()
   {
     location.reload();
+  }
+
+  openErrorDialog(errorMessage: string): void {
+    this.dialog.open(ErrorDialogComponent, {
+      data: { message: errorMessage }
+    });
   }
