+++ PedalProFrontEnd/PedalProFrontEnd/src/app/CompanyModule/training-module-company/training-module-company.component.ts
@@ -6,6 +6,9 @@ import { Router } from '@angular/router';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { map, Observable, Subject } from 'rxjs';
 import { NgModule } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent,MyDialogData } from 'src/app/Dialogs/delete-dialog/delete-dialog.component';
+import { ErrorDialogComponent } from 'src/app/Dialogs/error-dialog/error-dialog.component';
 
 @Component({
   selector: 'app-training-module-company',
@@ -13,7 +16,7 @@ import { NgModule } from '@angular/core';
   styleUrls: ['./training-module-company.component.css']
 })
 export class TrainingModuleCompanyComponent implements OnInit{
-  constructor(private service:PedalProServiceService,private router:Router, private http:HttpClient){
+  constructor(private dialog:MatDialog,private service:PedalProServiceService,private router:Router, private http:HttpClient){
   }
   statuses:ModuleStatus[]=[];
   modules:TrainingModule[]=[];
@@ -34,6 +37,12 @@ export class TrainingModuleCompanyComponent implements OnInit{
       });
     })
     return this.statuses;
+  }
+
+  openErrorDialog(errorMessage: string): void {
+    this.dialog.open(ErrorDialogComponent, {
+      data: { message: errorMessage }
+    });
   }
 
   GetModules(){
@@ -57,17 +66,36 @@ export class TrainingModuleCompanyComponent implements OnInit{
 
   DeleteModule(id:any)
   {
-    this.service.DeleteModule(id).subscribe({
-      next:(response)=>{
-        
-        const index=this.modules.findIndex((module)=>module.trainingModuleId===id);
-        if(index!=-1){
-          this.modules.slice(index,1);
-        }
-        this.openModal();
-        
+    const dialogData: MyDialogData = {
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this training module with all corresponding material?'
+    };
+
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.service.DeleteModule(id).subscribe({
+          next:(response)=>{
+            
+            const index=this.modules.findIndex((module)=>module.trainingModuleId===id);
+            if(index!=-1){
+              this.modules.slice(index,1);
+            }
+            this.openModal();
+            
+          },
+          error:(err)=>{
+            const errorMessage = err.error || 'An error occurred';
+            this.openErrorDialog(errorMessage);
+          }
+        })
       }
-    })
+    });
+    
+    
   }
 
   ReloadPage()
