+++ PedalProFrontEnd/PedalProFrontEnd/src/app/Timeslots/add-timeslot/viewtimeslots/viewtimeslots.component.ts
@@ -4,6 +4,9 @@ import { Router } from '@angular/router';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { map, Observable, Subject } from 'rxjs';
 import { DateWithTimeslotDto } from 'src/app/Models/date-with-timeslot-dto';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent,MyDialogData } from 'src/app/Dialogs/delete-dialog/delete-dialog.component';
+import { ErrorDialogComponent } from 'src/app/Dialogs/error-dialog/error-dialog.component';
 
 @Component({
   selector: 'app-viewtimeslots',
@@ -13,7 +16,7 @@ import { DateWithTimeslotDto } from 'src/app/Models/date-with-timeslot-dto';
 export class ViewtimeslotsComponent implements OnInit{
   timeslots:DateWithTimeslotDto[]=[];
 
-  constructor(private service:PedalProServiceService,private router:Router, private http:HttpClient){}
+  constructor(private dialog:MatDialog,private service:PedalProServiceService,private router:Router, private http:HttpClient){}
   ngOnInit(): void {
     this.GetBikeParts();
   }
@@ -31,17 +34,41 @@ export class ViewtimeslotsComponent implements OnInit{
   //Delete Bicycle
   DeleteBikePart(id:any)
   {
-    this.service.DeleteTimeslot(id).subscribe({
-      next:(response)=>{
-        
-        const index=this.timeslots.findIndex((bicyclePart)=>bicyclePart.timeslotId===id);
-        if(index!=-1){
-          this.timeslots.slice(index,1);
+    const dialogData: MyDialogData = {
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this timeslot?'
+    };
+
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.service.DeleteTimeslot(id).subscribe({
+          next:(response)=>{
+            
+            const index=this.timeslots.findIndex((bicyclePart)=>bicyclePart.timeslotId===id);
+            if(index!=-1){
+              this.timeslots.slice(index,1);
+            }
+            this.openModal();
+            
+          },
+          error:(err)=>{
+            const errorMessage = err.error || 'An error occurred';
+            this.openErrorDialog(errorMessage);
+          }
+        })
         }
-        this.openModal();
-        
-      }
-    })
+    });
+    
+  }
+
+  openErrorDialog(errorMessage: string): void {
+    this.dialog.open(ErrorDialogComponent, {
+      data: { message: errorMessage }
+    });
   }
 
   //Refresh the page
