+++ PedalProFrontEnd/PedalProFrontEnd/src/app/Workouts/WorkoutTypes/workout-type/workout-type.component.ts
@@ -4,6 +4,9 @@ import { Router } from '@angular/router';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { map, Observable, Subject } from 'rxjs';
 import { WorkoutType } from 'src/app/Models/workout-type';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent,MyDialogData } from 'src/app/Dialogs/delete-dialog/delete-dialog.component';
+import { ErrorDialogComponent } from 'src/app/Dialogs/error-dialog/error-dialog.component';
 
 @Component({
   selector: 'app-workout-type',
@@ -11,7 +14,7 @@ import { WorkoutType } from 'src/app/Models/workout-type';
   styleUrls: ['./workout-type.component.css']
 })
 export class WorkoutTypeComponent implements OnInit{
-  constructor(private service:PedalProServiceService,private router:Router, private http:HttpClient){}
+  constructor(private dialog:MatDialog,private service:PedalProServiceService,private router:Router, private http:HttpClient){}
 
   clientTypes:WorkoutType[]=[];
 
@@ -29,20 +32,46 @@ export class WorkoutTypeComponent implements OnInit{
     })
     return this.clientTypes;
   }
+
+  openErrorDialog(errorMessage: string): void {
+    this.dialog.open(ErrorDialogComponent, {
+      data: { message: errorMessage }
+    });
+  }
 // delete method
   DeleteEmpType(id:any)
   {
-    this.service.DeleteWorkoutType(id).subscribe({
-      next:(response)=>{
-        
-        const index=this.clientTypes.findIndex((clientType)=>clientType.workoutTypeId===id);
-        if(index!=-1){
-          this.clientTypes.slice(index,1);
-        }
-        this.openModal();
-        
+    
+    const dialogData: MyDialogData = {
+      title: 'Confirm Delete',
+      message: 'Are you sure you want to delete this workout type?'
+    };
+
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) { 
+        this.service.DeleteWorkoutType(id).subscribe({
+          next:(response)=>{
+            
+            const index=this.clientTypes.findIndex((clientType)=>clientType.workoutTypeId===id);
+            if(index!=-1){
+              this.clientTypes.slice(index,1);
+            }
+            this.openModal();
+            
+          },
+          error:(err)=>{
+            const errorMessage = err.error || 'An error occurred';
+            this.openErrorDialog(errorMessage);
+          }
+        })
       }
-    })
+    });
+    
+    
   }
 
   ReloadPage()
